+++ src/img/zimage2dutils.h
@@ -9,7 +9,6 @@
 #include "zimginterface.h"
 #include "zsaturateoperation.h"
 #include "zimagefilterkernel.h"
-#include <QList>
 #include <tbb/parallel_for.h>
 #include <boost/align/aligned_allocator.hpp>
 #include <algorithm>
@@ -265,7 +264,7 @@ void image2DPad(const TPixel* img, size_t width, size_t height,
       int refX = width;
       for (size_t i = leftPad; i-- > 0;) {
         --refX;
-        refX = refX < 0 ? width - 1 : refX;
+        refX = refX < 0 ? int(width) - 1 : refX;
         for (size_t j = 0; j < height; ++j) {
           imgOut[(j + upPad) * desWidth + i] = img[j * width + refX];
         }
@@ -283,14 +282,14 @@ void image2DPad(const TPixel* img, size_t width, size_t height,
       int refY = height;
       for (size_t j = upPad; j-- > 0;) {
         --refY;
-        refY = refY < 0 ? height - 1 : refY;
+        refY = refY < 0 ? int(height) - 1 : refY;
         std::memcpy(imgOut + j * desWidth,
                     imgOut + (refY + upPad) * desWidth,
                     sizeof(TPixel) * desWidth);
       }
       // down
       refY = -1;
-      for (int j = desHeight - downPad; j < static_cast<int>(desHeight); ++j) {
+      for (size_t j = desHeight - downPad; j < desHeight; ++j) {
         ++refY;
         refY = refY >= static_cast<int>(height) ? 0 : refY;
         std::memcpy(imgOut + j * desWidth,
